% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaBlack.R
\name{alphaBlack}
\alias{alphaBlack}
\title{Transparent Black}
\usage{
alphaBlack(alpha)
}
\arguments{
\item{alpha}{The level of opacity, between 0.0 and 1.0, 0.0 being completely
transparent and 1.0 being opaque.}
}
\value{
A character object of the hex value colour name created.
}
\description{
Create a black with a specified level of opacity.
}
\details{
The RGB values for alphaBlack are 0,0,0.
}
