% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{alphaBlue}
\alias{alphaBlue}
\title{Transparent Blue}
\usage{
alphaBlue(alpha)
}
\arguments{
\item{alpha}{The level of opacity, between 0.0 and 1.0, 0.0 being completely transparent and 1.0 being opaque.}
}
\value{
A character object of the hex value colour name created.
}
\description{
Create a blue with a specified level of opacity.
}
\details{
The RGB values used for alphaBlue are 0,0 and 128, where the maximum is 255.
}
