% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaGreen.R
\name{alphaGreen}
\alias{alphaGreen}
\title{Transparent Green}
\usage{
alphaGreen(alpha = 0.5)
}
\arguments{
\item{alpha}{The level of opacity, between 0.0 and 1.0, 0.0 being completely transparent and 1.0 being opaque. Default = 0.5.}
}
\value{
A character object of the hex value colour name created.
}
\description{
Create a green with a specified level of transparency.
}
\details{
The RGB values used for alphaGreen are 0,150 and 0, where the maximum is 255.
}
