% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaPink.R
\name{alphaPink}
\alias{alphaPink}
\title{Transparent Pink}
\usage{
alphaPink(alpha)
}
\arguments{
\item{alpha}{The level of opacity, between 0.0 and 1.0, 0.0 being completely transparent and 1.0 being opaque.}
}
\value{
A character object of the hex value colour name created.
}
\description{
Create a pink with a specified level of opacity.
}
\details{
The RGB values for alphaPink are 255,62 and 150, where the maximum is 255.
}
