% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaRed.R
\name{alphaRed}
\alias{alphaRed}
\title{Transparent Red}
\usage{
alphaRed(alpha)
}
\arguments{
\item{alpha}{The level of opacity, between 0.0 and 1.0, 0.0 being completely transparent and 1.0 being opaque.}
}
\value{
A character object of the hex value colour name created.
}
\description{
Create a red with a specified level of opacity.
}
\details{
The RGB values for alphaRed are 255,0 and 0, where the maximum is 255.
}
