% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis2Dmito-package.R
\docType{package}
\name{analysis2Dmito-package}
\alias{analysis2Dmito-package}
\alias{analysis2Dmito}
\title{The 'analysis2Dmito' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
