% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis2Dmito-package.R
\docType{package}
\name{analysis2Dmito}
\alias{analysis2Dmito-package}
\alias{analysis2Dmito}
\title{The 'analysis2Dmito packge'}
\description{
analysis2Dmito fits a Bayesian hierarchical linear mixture model, to
OXPHOS protein abundance data. The model is proposed in the paper
\verb{Bayesian classification of OXPHOS skeletal myofibres} by Childs et al.. The functions
provided here are specific for this purpose. Unforetunately, due to
lack of consistency in the structure of related data, tthe user must
first wrangle their data into a specified format. A detailed
description of the this and how to use the important fucntions within
the package can be found on the package
\href{https://github.com/jordanbchilds/analysis2Dmito}{GitHub page}.
}
\author{
\strong{Maintainer}: Jordan Childs \email{j.childs@newcastle.co.uk}

}
