% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcols.R
\name{classcols}
\alias{classcols}
\title{Classification Colour}
\usage{
classcols(classif, alphaLevel = 1)
}
\arguments{
\item{classif}{A numeric vector of deficiency probabilities, all elements must be between 0.0 and 1.0.}

\item{alphaLevel}{A number between 0.0 and 1.0 to control the level of the opacity of the colours. Default 1.0.}
}
\value{
A character vector with hex value RGB names.
}
\description{
The colour given to a fibre given its probability of being deficient. The
colours are on a scale from blue to red where pure blue represents a fibre
with a zero probability of being deficient and pure red represents a
deficiency probabilty of 1.0.
}
\details{
The colour gradient is created using \link{ramp_BlueRed}.
}
