% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData_doc.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example QIF dataset}
\format{
An object of class \code{data.frame} with 67575 rows and 5 columns.
}
\usage{
exampleData
}
\description{
This dataset has been taken from \href{https://www.medrxiv.org/content/10.1101/2022.04.04.22272484v1.full}{"Estimating the allele frequency threshold of
the pathogenic mtDNA variant m.3243A>G tolerated by human myofibres"} by Ahmed
et al. (2022).
}
\details{
The data is read from the \href{https://github.com/CnrLwlss/Ahmed_2022}{GitHub repository}
associated with the paper and so any changes made to the data there would
remain here. Preprocessing of the data here is only to remove values which
are less than or equal to zero. This so that the data can be logged
transformed without problem. Assuming the original dataset will not be
updated this results in the removal of two fibres.
}
\author{
Syeda T Ahmed, Robert W Taylor, Doug M Turnbull, Conor Lawless, Sarah T Pickett
}
\keyword{datasets}
