% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exampleData.R
\name{get_exampleData}
\alias{get_exampleData}
\title{Get Example Dataset}
\usage{
get_exampleData()
}
\value{
A \link{data.frame} object.
}
\description{
Returns an example data set. The dataset has been taken from
\href{https://www.medrxiv.org/content/10.1101/2022.04.04.22272484v1.full}{"Estimating t
he allele frequency threshold of the pathogenic mtDNA variant m.3243A>G
tolerated by human myofibres"} by Ahmed et al. (2022). It is read directly
the Github repository associated with the paper.

This dataset has been taken from \href{https://www.medrxiv.org/content/10.1101/2022.04.04.22272484v1.full}{"Estimating the allele frequency threshold of
the pathogenic mtDNA variant m.3243A>G tolerated by human myofibres"} by Ahmed
et al. (2022).
}
\details{
Preprocessing of the data here is only to remove values which are less than
or equal to zero. This so that the data can be logged transformed without
problem. Assuming the original dataset will not be updated this results in
the removal of two fibres.
}
\examples{
exampleData = get_exampleData()

}
\author{
Syeda T Ahmed, Robert W Taylor, Doug M Turnbull, Conor Lawless, Sarah T Pickett
}
