% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Retrive IMC or QIF Data}
\usage{
getData(fname, cord, correctnpc = TRUE, noNPCchannels = NULL)
}
\arguments{
\item{fname}{A character describing the filepath to the dataset from the current wokring directory.}

\item{cord}{A character vector of the channels of interest in the dataset.}

\item{correctnpc}{A boolean. If true the expression values are corrected NPC.}

\item{noNPCchannels}{If \code{correctnpc=TRUE} then the}
}
\value{
\code{NULL}
}
\description{
The function aggregates the data into the form needed for \code{\link[=inference]{inference()}}. To be able to do this the data passed to the function must in a specific form, see details for more info.
The data.frame object passed to the function must be in long form and have the following columns; 'value', 'channel', 'sampleID' and 'fibreID'. Where 'value' is the protein expression value, 'channel' is the protein or channel that the value expresses, 'sampleID' is the identifying name associated with the sample the expression is from and'fibreID' is the fibre identification from that sample. Other columns can be present but are not needed.
To be able to transform the data into long form, we suggest using the \link{tidyr} package and the \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} function.
}
\details{
The file is read using the \link[data.table:fread]{data.table::fread} function. If the file does not
exist here then an error is returned stating as such.

For NPC correction to take place the \code{filenames} column in the dataset must
be apprropriately labelled as "\if{html}{\out{<sampleID>}} NPC" or "\if{html}{\out{<sampleID>}} OXPHOS"
}
