% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_saver.R
\name{list_saver}
\alias{list_saver}
\title{Save List of Dataframes}
\usage{
list_saver(
  df_list,
  root,
  replace = TRUE,
  sep = "\\t",
  col.names = TRUE,
  row.names = FALSE,
  ...
)
}
\arguments{
\item{df_list}{A list of dataframes to be saved.}

\item{root}{A character vector of the root paths to which the names of the list will be added. If this is a single character then the same root will be used for files.}

\item{replace}{A boolean variable. If TRUE the function will overwrite any file that exists under the same name.}

\item{sep}{A character of an appropriate separator value.}

\item{col.names, row.names}{A boolean, passed directly to \code{\link[=write.table]{write.table()}}.}

\item{...}{Any additional parameters to be passed to write.table.}
}
\description{
Saves all data frames stored within a list by suffixing the name of the
dataframe, as written in the list,  onto the end of a root file path. The
files are saved as .txt files defaulted to the tab separator.
}
\details{
The function uses the \code{\link[=write.table]{write.table()}} function to save the files.
}
\examples{
data_list = list(fileOne = data.frame(a=1:20, b=21:40), fileTwo=data.frame(x=1:1000, y=rnorm(1000)))
dir.create("dir")
rootFile = "data/sourceOne"
output_saver(data_list, rootFile, replace=FALSE)

}
