% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp_BlueRed.R
\name{ramp_BlueRed}
\alias{ramp_BlueRed}
\title{Blue-Red Colour Ramp}
\usage{
ramp_BlueRed(alphaLevel)
}
\arguments{
\item{alphaLevel}{The level of opacity of the two end-point colours. A number between 0.0 and 1.0, where 0.0 is completely transparent and 1.0 being opaque.}
}
\value{
A character object of the hex value colour name created.
}
\description{
For creating transitions colours between red and blue, with a given transparency.
}
\details{
The colour gradient lies between \code{\link[=alphaBlue]{alphaBlue()}} and \code{\link[=alphaRed]{alphaRed()}}
}
