% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{alphaCol}
\alias{alphaCol}
\title{Transparent Colour Creator}
\usage{
alphaCol(alpha, name = NULL, rgb = NULL, maxValue = 1)
}
\arguments{
\item{alpha}{The level of transparency}

\item{name}{A valid colour name, as listed by \code{\link[=colors]{colors()}}.}

\item{rgb}{A numeic vector of RGB values.}

\item{maxValue}{The maximum colour value, default to 1.}
}
\description{
Defining a transparent version of a given colour.
}
