% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exampleData.R
\name{get_exampleData}
\alias{get_exampleData}
\title{Get Example Dataset}
\usage{
get_exampleData()
}
\value{
A \link{data.frame} object.
}
\description{
Returns an example data set. The dataset consists of protein expression levels
collected via IMC, for VDAC, NDUFB8, CYB and MTCO1, on 12 patients.

The data is returned as a \link{data.frame} object in the correct form required for
use in the \link{getData_mats} function.
}
\examples{
exampleData = get_exampleData()

}
\author{
Amy Vincent et al.
}
