% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Bayesian Inference for a Hierarchical Linear Mixture Model with a Bernoulli Classifier}
\usage{
inference(
  dataMats,
  parameterVals = NULL,
  MCMCout = 1000,
  MCMCburnin = 1000,
  MCMCthin = 1,
  MCMCchain = 1
)
}
\description{
Using \link{rjags} the function fits a hierarchical linear mixture model to a
dataset. The was developed to be able to classify individual fibres within a
2Dmito plot as being a set of data control fibre datapoints. The model
fundamentally assumes the contrl data shows strong linearity and anything
diverges from this is not like control.
}
